requirements.txt
sqlalchemy==2.0.35
pydantic==2.9.2
passlib==1.7.4
databases==0.9.0
fastapi==0.115.0
uvicorn==0.30.6
asyncpg==0.30.0
psycopg2==2.9.9
python-dotenv==1.0.1
python-jose==3.3.0